Extract boundaries

Returns boundary geometries of the currently mounted extract. For local instances, get_extract_boundaries wraps oe_read. For remote instances, see section "Remote instances".

Usage

get_extract_boundaries(instance = NULL, force = FALSE, verbose = TRUE, ...)

Arguments

instance

[ors_instance]

Object of an OpenRouteService instance that should be used for route computations. It is recommended to use ors_instance to set an instance globally. This argument should only be used if activating an instance globally is not feasible.

force

[logical]

If TRUE, extract must be identified and parsed. If FALSE, the geometries will be read from the cache if possible.

verbose

[logical]

If TRUE, prints a loading spinner.

...

If instance is a local instance, this is ignored. Otherwise, specifies further arguments for generating a random bbox including: dist, the bbox radius in meters, and seed, a random seed as in set.seed.

Value

An sfc object of the currently mounted extract boundaries.

Remote instances

If instance is not local, it is more difficult to derive the extract boundaries. There is thus far no way of accessing an OSM extract file knowing only the server address. We can, however, make use of some heuristics:

ors_extract can export the built graphs from an ORS server if it allows it. However, it does not work on the public API and it requires knowledge about the approximate area of an extract.

ors_guess can make an approximation of an extract area. It accesses the snap endpoint which also does not work on the public API and needs to be enabled on other servers. ors_guess can make a lot of requests and might not be feasible in many situations.

Examples

if (FALSE) {
# For local instances, reads the extract file
bounds <- get_extract_boundaries()
plot(bounds)

# Subsequent calls recover from the cache
get_extract_boundaries()
}