ORS accessibility maps

ors_accessibility returns isochrone or isodistance data for a set of source points, either as polygons or as a raster.

Usage

ors_accessibility(
  src,
  profile = get_profiles(force = TRUE),
  range = c(200L, 300L),
  attributes = "area",
  intersections = FALSE,
  interval = 30L,
  location_type = c("start", "destination"),
  range_type = c("time", "distance"),
  smoothing = 25L,
  area_units = c("m", "km", "mi"),
  units = c("m", "km", "mi"),
  rasterize = FALSE,
  raster_resolution = c(100L, 100L),
  instance = NULL,
  ...
)

Arguments

src

[sf/sfc]

Source dataset containing point geometries that should be routed from.

profile

[character]

Means of transport as supported by OpenRouteService. Defaults to the first profile in a call to get_profiles. For ors_shortest_distances, profile can be a character vector, for all other functions it needs to be a character scalar. For details on all profiles, refer to the documentation.

range

[character]

If length-1, specifies the maximum range which can be further broken down using the argument interval. If of length larger than 1, specifies manual breaks. The unit can be controlled using the argument range_type.

attributes

[character]

Additional features to be included the output. Only relevant if rasterize = FALSE. Can be one of "area", "reachfactor" and "total_pop".

intersections

[logical]

Wether to return overlapping polygons.

interval

[integer]

If range is a length-1 vector, defines the distance interval to break down the isochrones.

location_type

[character]

Whether to route from ("start") or to ("destination") the points of the input data.

range_type

[character]

Type of distance that the calculations should be based on. "time" produces isochrones while "distance" produces isodistance polygons.

smoothing

[integer]

Smoothing factor between 0 and 100 to be applied to the polygon geometries. For details, refer to the API playground.

area_units

[character]

Distance unit to use for the area attribute from attributes.

units

[character]

Units to be used for the range argument if range_type = "distance".

rasterize

[logical]

If FALSE, returns the isochrone polygon geometries. If TRUE, aggregates polygon distance values onto a vector grid and then rasterizes the vector grid to a SpatRaster object. When rasterizing, only the distance value is preserved while all variables specified in attributes are discarded.

raster_resolution

[numeric]

If rasterize = TRUE, specifies the resolution of the template raster. Corresponds to the number of cells in each direction (x, y). This argument is passed on to st_make_grid and rast.

instance

[ors_instance]

Object of an OpenRouteService instance that should be used for route computations. It is recommended to use ors_instance to set an instance globally. This argument should only be used if activating an instance globally is not feasible.

...

Additional arguments passed to the ORS API. This includes all options that modify the routing results. For details on each argument, refer to the API playground and documentation.

avoid_borders

Length-1 character vector specifying whether to avoid, all borders, only controlled ones or none. Only available for driving-*.

avoid_countries

Numeric vector listing countries to avoid. Each country is assigned a numeric value. Refer to the ORS documentation. Only available for driving-*.

avoid_features

Character vector containing traffic features to avoid (e.g. highways or tunnels).

avoid_polygons

sf or sfc object describing areas to avoid.

profile_params

Nested list containing restrictions and weightings for driving-hgv, cycling-*, walking, hiking and wheelchair.

vehicle_type

Length-1 character vector specifying the type of heavy goods vehicle. Needed to set restrictions for driving-hgv.

Value

If rasterize = FALSE, returns an sf object containing the isochrone or isodistance polygon geometries as well as additional attributes. Each polygon's outer lines represent places of equal distance or travel time from or to the respective point in the source dataset. If rasterize = FALSE, returns a raster object with a discrete value classification.

Examples

if (FALSE) {
set.seed(123)
src <- ors_sample(10)
data("pharma")

# Returns a polygon sf dataframe divided by four 15 minute time breaks
# Also contains information on the area and population inside the isochrones
ors_accessibility(
  source_sf,
  profile = "driving-car",
  range = c(900, 1800, 2700, 3600),
  attributes = c("area", "total_pop"),
  location_type = "start",
  range_type = "time",
  rasterize = TRUE
)

# Returns a raster image classified by 3 distance intervals
ors_accessibility(
  source_df,
  profile = "driving-car",
  range = c(1000, 5000, 10000),
  location_type = "destination",
  range_type = "distance"
)

# Returns a polygon sf dataframe that is broken down every 50 meters up until
# the maximum distance of 500 meters.
ors_accessibility(
  source_df,
  profile = "cycling-regular",
  range = 500,
  location_type = "start",
  interval = 50,
  range_type = "distance"
)}