OpenRouteService instance

Creates an OpenRouteService instance object. An instance represents either a local/remote server or a directory from which OpenRouteService can be set up. Running this function or any of the related functions listed below stores the instance in an internal environment object and enables functions like ors_pairwise to automatically detect the appropriate server information needed to make a successful request. Hence, this function should always be run after loading ORSRouting as a means of fixing an instance to the current session.

While initializing an instance using an already running server requires no further action, this function family excels at building a local OpenRouteService server from source. Setting up a local server effectively removes any server-side rate limits and allows you to conveniently use the package functions on much larger datasets. For setting up a local server, it is required to build and start a Docker container. To do this, ors_instance starts Docker (if necessary), downloads and unpacks the OpenRouteService source code and returns an object of class ors_instance.

For a full reference of the methods of an ORS instance, refer to ORSLocal and ORSRemote.

Usage

ors_instance(
  dir = "~",
  server = NULL,
  version = "latest",
  overwrite = FALSE,
  verbose = 2L,
  ...
)

Arguments

dir

[character]

Custom OpenRouteService directory. If not specified, the directory will be downloaded to the system's home directory. If a directory called "openrouteservice-{version}" is present, the download will be skipped. Ignored if server is not NULL.

server

[character]

URL of a server that accepts OpenRouteService requests. This can be a URL to a local or a remote server. The official public API can be accessed using the shortcut "public". Keep in mind that the public API is rate-restricted and requests are automatically throttled to 40 requests per minute. Routing functions will be slow for larger datasets.

version

[character]

The OpenRouteService version to use. Can either be a version number (e.g. 7.2.0), a commit hash, or "master". Ignored if server is not NULL.

overwrite

[logical]

Whether to overwrite the current OpenRouteService directory if it exists.

verbose

[logical]

If TRUE, prints informative messages and spinners.

...

Further arguments passed to ORSLocal or ORSRemote.

Value

R6 object of class ors_instance as created by ORSInstance

Examples

if (FALSE) {
dir <- dir.create("~/test_ors")

# Download and furnish an ORS instance
ors_instance(dir = dir, version = "7.2.0")

# Connect to the public API
ors_instance(server = "public")

# Connect to a local server
ors_instance(server = "https://127.0.0.1:8001/")
}