Routing distance computations

ors_pairwise calculates the routing distance between two datasets using the Directions service from ORS. ors_shortest_distances is a wrapper around ors_pairwise that matches each point of the source dataset to a dataset of points of interest from the destination dataset and then extracts the route with the shortest distance.

Usage

ors_pairwise(
  src,
  dst,
  profile = get_profiles(),
  units = c("m", "km", "mi"),
  geometry = FALSE,
  instance = NULL,
  ...
)

ors_shortest_distances(
  src,
  dst,
  group = NULL,
  profile = get_profiles(),
  units = c("m", "km", "mi"),
  geometry = FALSE,
  instance = NULL,
  ...,
  proximity_type = c("duration", "distance"),
  progress = TRUE
)

Arguments

src

[sf/sfc]

Source dataset containing point geometries that should be routed from.

dst

[sf/sfc]

Destination dataset containing point geometries that should be routed to. For ors_shortest_distances, the destination argument can also be a dataframe containing a grouping column specified by the group argument that indicates which destinations refer to which row in the source dataset (as returned by get_closest_pois). This is recommended for large datasets because passing a plain sf dataframe routes from each source point to each point in the entire destination dataset.

profile

[character]

Means of transport as supported by OpenRouteService. Defaults to the first profile in a call to get_profiles. For ors_shortest_distances, profile can be a character vector, for all other functions it needs to be a character scalar. For details on all profiles, refer to the documentation.

units

[character]

Distance unit for distance calculations ("m", "km" or "mi")

geometry

[logical]

If TRUE, returns a sf object containing route geometries. If FALSE, returns route distance measures. Defaults to FALSE, to increase performance.

instance

[ors_instance]

Object of an OpenRouteService instance that should be used for route computations. It is recommended to use ors_instance to set an instance globally. This argument should only be used if activating an instance globally is not feasible.

...

Additional arguments passed to the ORS API. This includes all options that modify the routing results. For details on each argument, refer to the API playground and documentation

geometry_simplify

Whether geometry should be simplified. Defaults to FALSE.

continue_straight

Whether to avoid u-turns and continue straight if possible. If FALSE, allows the route to take u-turns. Defaults to FALSE.

avoid_borders

Character vector specifying whether to avoid, all borders, only controlled ones or none. Only available for driving-*.

avoid_countries

Numeric vector listing countries to avoid. Each country is assigned a numeric value. Refer to the ORS country list. Only available for driving-*.

avoid_features

Character vector containing traffic features to avoid. One or several of highways, tollways, ferries, fords, and steps.

avoid_polygons

sf or sfc object describing areas to avoid.

restrictions

Restrictions to be applied for routing. Routes that do not conform to these restrictions are not considered. Can either be specifications for heavy-goods vehicles (axleload, hazmat, height, length, weight, width), indicators of wheelchair accessibility (maximum_incline, maximum_sloped_kerb, minimum_width, smoothness_type, surface_type, track_type), or the maximum allowed route steepness (gradient). Not applicable for any other profile. Refer to the documentation for details on each parameter.

weightings

Weightings for route selection. Can be preference multiplicators for walking profiles (green, quiet) or steepness_difficulty for cycling profiles. Not applicable for any other profile. Refer to the documentation for details on each parameter.

allow_unsuitable

Whether possibly unsuitable surfaces should be included for wheelchair routing. If TRUE, includes routes, even if their surface quality might not be suitable for wheelchair driving. Defaults to FALSE.

surface_quality_known

Whether to enforce that the surface quality of routes is considered. If FALSE, routes of all surface qualities are taken into account. Defaults to FALSE

vehicle_type

The type of heavy-goods vehicle that is to be assumed for profile driving-hgv. One of hgv, agricultural, delivery, forestry, and goods. Needed to set restrictions for driving-hgv. Defaults to hgv.

preference

Specifies the route preference mode. One of recommended, fastest or shortest. Defaults to recommended.

radiuses

Maximum distance (in m) that road segments can be snapped to. radiuses = -1 represents an unlimited radius. Defaults to the maximum_snapping_radius in the configuration file (350m if not changed).

maximum_speed

Maximum speed that routing vehicles are allowed to drive. Not applied by default.

group

[character/numeric]

Column name or index providing a grouping column that indicates which row in the destination dataset corresponds to which row in the source dataset (as in the output of get_closest_pois). Providing a grouping column can considerably reduce the processing load for larger datasets.

proximity_type

[character]

Type of proximity that the calculations should be based on. If distance, the shortest physical distance will be calculated and if duration, the shortest temporal distance will be calculated.

progress

[logical]

If TRUE, displays a progress bar if the process is taking a bit longer.

Value

ors_pairwise returns a dataframe with distances and travel durations between source and destination. ors_shortest_distances returns a dataframe containing distances, travel durations and the index number of the point of interest with the shortest routing distance to the respective place of the source dataset. Depending on the geometry argument, these outputs can either be simple dataframes or objects of class sf containing the linestring geometries of the respective routes.

Details

For ors_pairwise, the profile argument supports only length-1 vectors while ors_shortest_distances supports multiple profiles. ors_shortest_distances finds the shortest route for each source point and each profile, respectively.

Error handling

Since ors_pairwise is supposed to conduct a lot of calculations in one go, errors might occur even in well-conceived service setups. In order to make debugging less painful, errors do not tear down the whole process. They are saved to an environment and issue a warning containing the indices of the routes in question. After the process has finished, they can be accessed by calling last_ors_conditions. Specific routes can be examined by inspecting its route attributes using ors_inspect.

Examples

if (FALSE) {
data("pharma")

set.seed(123)
dest <- ors_sample(10)

car <- "driving-car"
bike <- "cycling-regular"

# Running with sf objects
route_lengths_sf <- ors_pairwise(pharma, dest, profile = car)
route_lengths_sf

# Running with coordinate pairs
route_lengths_df <- ors_pairwise(pharma, dest, profile = bike)
route_lengths_df

# Returns route geometries
route_lengths_geom <- ors_pairwise(
  pharma,
  dest,
  profile = car,
  geometry = TRUE
)

# Returns routes in kilometers
route_lengths_km <- ors_pairwise(
  pharma,
  dest,
  profile = bike,
  units = "km"
)

# Running with additional arguments
route_lengths_opts <- ors_pairwise(
  pharma,
  dest,
  profile = car,
  continue_straight = TRUE,
  preference = "fastest"
)

# Finding shortest routes from each point in sample_a to sample_b
shortest_routes <- ors_shortest_distances(pharma, dest, units = "km")
shortest_routes

# Pre-filter the nearest 5 destination points by Euclidian distance
pois <- get_closest_pois(pharma, dest, n = 5)

# Only route from each pharmacy to one of the closest 5 destination points
# respectively. For larger datasets, this can increase performance.
nearest_hospitals <- ors_shortest_distances(
  pharma,
  pois,
  group = ".group",
  geometry = TRUE
)
nearest_hospitals
}
if (FALSE) {

}